+++ src/main/java/com/lmsf/org/service/BookService.java
@@ -94,6 +94,8 @@ public class BookService {
         return bookRepository.save(book);
     }
     public void deleteBook(Long id){
+        if(!bookRepository.existsById(id))
+            throw new BookNotFoundException("Book not found with id : "+id);
         bookRepository.deleteById(id);
     }
     public Set<Genre> getGenres(Long id) {
