+++ src/main/java/com/lmsf/org/service/GenreService.java
@@ -3,6 +3,7 @@ package com.lmsf.org.service;
 import com.lmsf.org.dto.GenreDto;
 import com.lmsf.org.entity.Book;
 import com.lmsf.org.entity.Genre;
+import com.lmsf.org.exception.GenreDeleteException;
 import com.lmsf.org.exception.GenreNotFoundException;
 import com.lmsf.org.repository.BookRepository;
 import com.lmsf.org.repository.GenreRepository;
@@ -24,7 +25,8 @@ public class GenreService {
     }
     public void deleteGenre(Long id){
         List<Book> books = bookRepository.findByGenresId(id);
-        books.forEach(book -> bookRepository.deleteById(book.getId()));
+        if(!books.isEmpty())
+            throw new GenreDeleteException("Cannot delete the genre with id '" + id + "' because it is associated with " + books.size() + " books.");
         genreRepository.deleteById(id);
     }
     public Genre updateGenre(Long id, Genre genre){
